pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "cx-oracle>=8.3.0",
    "email-validator>=2.2.0",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "google-cloud-language>=2.17.0",
    "gunicorn>=23.0.0",
    "mysql-connector-python>=9.2.0",
    "numpy>=2.2.3",
    "pandas>=2.2.3",
    "psycopg2-binary>=2.9.10",
    "pyodbc>=5.2.0",
    "sift-stack-py>=0.4.2",
    "sqlalchemy>=2.0.39",
]
